adds default dir, exception for no dir, moves tests
